+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 - Custom git commit message template (@mszostak)
 - Redirect user into `/` after checkin login failure (@mkasztelnik)
 - Service categories (@mkasztelnik)
+- Service order possibility (@mkasztelnik)
 
 ### Changed